crd: add new attributes to the status sub-resource
This commits adds attributes to the status sub-resource:
- status.Conditions: controller can report success / errors while
  reserving IP addresses for a workload
- status.OwnerPod: controller reports which pod owns the allocation

Adding conditions to the status is used to inform the user of allocation
errors (i.e. no IP addresses available in the subnet, the requested IP
addr was already in use, etc) while adding a reference to which pod owns the
allocation is used to provide information to the user, and to ensure that a
claim in use cannot be used by another workload.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>